Refactor B-spline computation to dedicated math module
This commit extracts the B-spline logic from templates.py into a new src/furax/math/bspline.py module. It also optimizes the HWP spline basis generation, exposes the bspline module in furax.math, and updates the relevant tests.
