+++ src/furax/math/__init__.py
@@ -1,0 +1,3 @@
+from . import bspline, quaternion, sht
+
+__all__ = ['bspline', 'quaternion', 'sht']
