+++ tests/mapmaking/test_mapmaker.py
@@ -385,7 +385,7 @@ class TestSingleObsTemplates:
         # amplitudes are structured (n_dets, *basis_shape), not flat
         assert res['template_polynomial'].shape == (n_dets, 2, 4)  # intervals x (orders 0..3)
         assert res['template_azhwp_synchronous'].shape == (n_dets, 4, 9)  # orders x (DC+2*4)
-        assert res['template_spline_hwpss'].shape == (n_dets, 2 * (20 + 2))  # 2 * (n_knots + 2)
+        assert res['template_spline_hwpss'].shape == (n_dets, 2 * (100 + 2))  # 2 * (n_knots + 2)
         for key, value in res.items():
             if key.startswith('template_') and not key.startswith('template_reg'):
                 assert bool(jnp.all(jnp.isfinite(value))), key
